+++ Graphic11/base/texture.cpp
@@ -347,7 +347,7 @@ string cTexture::ConvertTextureFileName(const char *fileName)
 	const Str32 ext = out.GetFileExt();
 	if ((ext == ".tga") || (ext == ".TGA"))
 	{
-		const int len = out.size();
+		const size_t len = out.size();
 		out.m_str[len - 3] = 'd';
 		out.m_str[len - 2] = 'd';
 		out.m_str[len - 1] = 's';
